pyproject.toml
[project]
name = "fastapi-study"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.0.1",
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "uvicorn>=0.42.0",
]
